% adds stats to getProperties; % changes currencies to EUR; % adds image to items; % adds position to PropertyDTO and DAO; % sorts list before showing it.

+++ network/src/main/kotlin/com/lc/magicproperties/network/ApiService.kt
@@ -14,7 +14,7 @@ class ApiService(httpClientFactory: IHttpClientFactory<IApiRetrofitServiceDefini
         return api.getProperties()
     }
 
-    override fun getStats(action: String, duration: Int): Single<Response<Int>> {
+    override fun getStats(action: String, duration: Long): Single<Response<Int>> {
         return api.getStats(action, duration)
     }
